fix : 출고의 개수가 아닌, 출고된 상품의 개수를 조회하는 방식으로 변경

+++ SunInCloud/src/main/java/com/shinhan/sunInCloud/service/ExportsService.java
@@ -205,7 +205,7 @@ public class ExportsService {
 		}
 		for (Object[] count : counts) {
 			String date = String.valueOf(count[0]);
-			BigInteger bi = (BigInteger) count[1];
+			BigDecimal bi = (BigDecimal) count[1];
 			map.put(date, bi.longValue());
 		}
 		for (String date : dates) {
